+++ CHANGELOG.md
@@ -13,6 +13,7 @@ details.
 
 ## [Unreleased]
 ### Added
+- checkrules.js: lookup `ESLint` Github repository for new/removed rules.
 - All published rules as of ESLint 2.2.0
 
 ### Changed